Remove visual regression tests
Since jest snapshots now contain CSS we can reliably recreate a visual
representation of a component on-demand, making the screenshot testing
redundant.
We will need to increase the coverage of our snapshot testing to cover
the variants in storybook if we are to achieve the same test coverage as
the screenshot testing and this will be done in a follow-up commit.

Change-type: minor
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>